py/manipulation/requirements_external.txt
absl-py == 0.9.0
numpy >= 1.16.0
typing-extensions >= 3.10.0.0
